refactor(scan): move buffer extractors into a discovery module
The four extractors and their helpers move verbatim into discovery.lua,
where they become hooks taking a context rather than loose arguments,
and are run through a chain that mirrors the matcher chain. The path
helpers move to path.lua, which both modules use, so discovery does not
depend on the scanner that depends on it.

Nothing about the behavior changes: the shipped priorities reproduce
the order the extractors ran in, all four sit under '*' because buffer
discovery has always been filetype agnostic, and GAPDoc keeps appending
its own extension through the new extension flag. The bodies are byte
identical to what they replace, apart from the path helper calls being
renamed. find_bib_files_from_buffer gains an optional opts argument and
returns exactly what it did before.

The suite is unchanged and green, which is what this commit is for.

Ref: #27
